Fixed and tested mumble context for different cells
The problem was that if 2 players were in a different cell, then their context would be different. As far as I know, mumble handles this by treating those 2 players as if they weren't playing the same game, so they would hear eachother at regular volume.

Not only is this annoying when some players are indoors and others are outside, but on the border of cells (such as balmora to West Gash) this would happen, even if the two characters were very close.

Setting the context only to the server's network address fixes both of these problems. Interiors and the World have a different different coordinate system, so the player's interior position is (usually) far enough away from their world position. Thus players indoors will not hear players outside. Furthermore, this allows players on cell borders to hear eachother as expect.