Fix all interiors being near eachother
There are interior cells that are intuitively supposed to be near each other but are not within audible range due. There are also some cases in which interiors are in the same instance, but should not logically be. In this case 2 players in each interior would hear each other, which is not desirable.

This change hashs the interior cell name, then converts that to a float, which is used as the offset for Mumble's actor coordinates. Thus all Mumble actor coordinates will be (offset + x, offset + z, offset + y).